* Make Phan fully static * Split Phan supplementary responsibilities * Minor fixes
