+++ src/Phan/Analyze/ParentClassExistsAnalyzer.php
@@ -7,7 +7,7 @@ use \Phan\Language\Element\Clazz;
 use \Phan\Language\FQSEN;
 use \Phan\Log;
 
-trait ParentClassExists {
+class ParentClassExistsAnalyzer {
 
     /**
      * Check to see if the given Clazz is a duplicate
