+++ src/Phan/Analyze/PropertyTypesAnalyzer.php
@@ -1,23 +1,19 @@
 <?php declare(strict_types=1);
 namespace Phan\Analyze;
 
-use \Phan\CodeBase;
-use \Phan\Issue;
-use \Phan\Language\Element\Clazz;
-use \Phan\Language\FQSEN;
-use \Phan\Log;
+use Phan\CodeBase;
+use Phan\Issue;
+use Phan\Language\Element\Clazz;
+use Phan\Language\FQSEN;
 
-trait PropertyTypes {
+class PropertyTypesAnalyzer {
 
     /**
      * Check to see if the given Clazz is a duplicate
      *
      * @return null
      */
-    public static function analyzePropertyTypes(
-        CodeBase $code_base,
-        Clazz $clazz
-    ) {
+    public static function analyzePropertyTypes(CodeBase $code_base, Clazz $clazz) {
         foreach ($clazz->getPropertyList($code_base) as $property) {
             $union_type = $property->getUnionType();
