+++ tests/Phan/AnalyzerTest.php
@@ -8,7 +8,7 @@ $internal_interface_name_list = get_declared_interfaces();
 $internal_trait_name_list = get_declared_traits();
 $internal_function_name_list = get_defined_functions()['internal'];
 
-use \Phan\Phan;
+use \Phan\AstAnalyzer;
 use \Phan\CodeBase;
 use \Phan\Config;
 use \Phan\Language\Context;
@@ -60,7 +60,7 @@ class AnalyzerTest extends \PHPUnit_Framework_TestCase {
                 Class A {}
             ");
 
-        $this->assertTrue(
+        self::assertTrue(
             $this->code_base->hasClassWithFQSEN(
                 FullyQualifiedClassName::fromFullyQualifiedString('A')
             )
@@ -74,7 +74,7 @@ class AnalyzerTest extends \PHPUnit_Framework_TestCase {
                 Class B {}
             ");
 
-        $this->assertTrue(
+        self::assertTrue(
             $this->code_base->hasClassWithFQSEN(
                 FullyQualifiedClassName::fromFullyQualifiedString('\A\b')
             )
@@ -95,7 +95,7 @@ class AnalyzerTest extends \PHPUnit_Framework_TestCase {
         $class_fqsen =
             FullyQualifiedClassName::fromFullyQualifiedString('\A\b');
 
-        $this->assertTrue(
+        self::assertTrue(
             $this->code_base->hasClassWithFQSEN($class_fqsen),
             "Class with FQSEN $class_fqsen not found"
         );
@@ -103,7 +103,7 @@ class AnalyzerTest extends \PHPUnit_Framework_TestCase {
         $clazz =
             $this->code_base->getClassByFQSEN($class_fqsen);
 
-        $this->assertTrue(
+        self::assertTrue(
             $clazz->hasMethodWithName($this->code_base, 'c'),
             "Method with FQSEN not found"
         );
@@ -118,7 +118,7 @@ class AnalyzerTest extends \PHPUnit_Framework_TestCase {
     ) : Context {
 
         return
-            (new Phan)->parseNodeInContext(
+            AstAnalyzer::parseNodeInContext(
                 $this->code_base,
                 new Context,
                 \ast\parse_code(
